+++ opentelemetry-sdk/src/trace/sampler/jaeger_remote/mod.rs
@@ -1,0 +1,10 @@
+mod rate_limit;
+mod remote;
+mod sampler;
+mod sampling_strategy;
+
+pub(crate) use sampler::JaegerRemoteSampler;
+pub use sampler::JaegerRemoteSamplerBuilder;
+
+#[cfg(test)]
+mod tests {}