+++ opentelemetry-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## to be released
+
+### Changed
+
+- *BREAKING* `struct`s which implement `ShouldSample` a.k.a Custom Samplers must now
+  implement `Clone`. This enables (#833)
+
 ## v0.1.0
 
 - SDK split from `opentelemetry` crate